Crate sel4_task

This crate provides task management for seL4, including the TCB, scheduler, and thread relevant structures.

See more details in ../doc.md

Re-exports

pub use sync::UPSafeCell;

pub use tcb::*;

Modules

heap

sync

tcb

Structs

dschedule_t

Struct representing a domain schedule.

lookupSlot_raw_ret_t

Structure for the return value of lookupSlot_raw

lookupSlot_ret_t

Structure for the return value of lookupSlot

tcb_queue_t

Structure for the tcb queue

thread_state_t

Enums

ThreadState

The state of a thread

Constants

SchedulerAction_ChooseNewThread

SchedulerAction_ResumeCurrentThread

ksDomScheduleLength

Statics

ksCurDomain

ksCurThread

ksDomSchedule

ksDomScheduleIdx

ksDomainTime

ksIdleThread

ksReadyQueues

ksReadyQueuesL1Bitmap

ksReadyQueuesL2Bitmap

ksSchedulerAction

ksWorkUnitsCompleted

Functions

activateThread

Activate the current thread.

addToBitmap

Add the given priority level to the ready queue bitmap.

create_idle_thread

Create the idle thread.

get_currenct_thread

Get the current thread, and returns a mutable tcb reference to the current thread. FIXME: fix the name of this function, get_current_thread

get_currenct_thread_unsafe

Get the current thread, and returns a mutable tcb reference to the current thread unsafely.

get_current_domain

Get the current domain.

get_idle_thread

Get the idle thread, and returns a mutable tcb reference to the idle thread.

get_ks_scheduler_action

Get the action to be taken by ks scheduler.

idle_thread

isHighestPrio

Check if the given priority level is the highest priority level for the given domain.

possible_switch_to

Schedule the given tcb when current tcb is not in the same domain or current action is not to resume the current thread.

ready_queues_index

Get the index of the ready queue for the given domain and priority level.

removeFromBitmap

Remove the given priority level from the ready queue bitmap.

rescheduleRequired

Reschedule threads, and enqueue the current thread if current ks scheduler action is not to resume the current thread and choose new thread.

schedule

Schedule threads.

schedule_tcb

Schedule the given tcb.

set_current_scheduler_action

Set the action to be taken by current scheduler.

set_current_thread

Set the current thread.

set_ks_scheduler_action

Set the action to be taken by ks scheduler.

timerTick

Schedule current thread if time slice is expired.