Struct sel4_task::tcb::tcb_t

#[repr(C)]
pub struct tcb_t {

    pub tcbArch: ArchTCB,
    pub tcbState: thread_state_t,
    pub tcbBoundNotification: usize,
    pub tcbFault: seL4_Fault_t,
    pub tcbLookupFailure: lookup_fault_t,
    pub domain: usize,
    pub tcbMCP: usize,
    pub tcbPriority: usize,
    pub tcbTimeSlice: usize,
    pub tcbFaultHandler: usize,
    pub tcbIPCBuffer: usize,
    pub tcbSchedNext: usize,
    pub tcbSchedPrev: usize,
    pub tcbEPNext: usize,
    pub tcbEPPrev: usize,

}

Structure for the TCB

Fields
tcbArch: ArchTCB

The architecture registers of the TCB

tcbState: thread_state_t

The state of the TCB

tcbBoundNotification: usize

The bound notification of the TCB

tcbFault: seL4_Fault_t

The fault of the TCB

tcbLookupFailure: lookup_fault_t

The lookup fault of the TCB

domain: usize

The domain of the TCB

tcbMCP: usize

The maximum controlled priority of the TCB

tcbPriority: usize

The priority of the TCB

tcbTimeSlice: usize

The time slice of the TCB

tcbFaultHandler: usize

The falut handler of the TCB

tcbIPCBuffer: usize

The IPC buffer of the TCB

tcbSchedNext: usize

The next TCB in the scheduling queue

tcbSchedPrev: usize

The previous TCB in the scheduling queue

tcbEPNext: usize

The next TCB in the EP queue

tcbEPPrev: usize

The previous TCB in the EP queue

Implementations

impl tcb_t

pub fn get_cspace(&self, i: usize) -> &'static cte_t

Get i th cspace of the TCB, unmutable reference

Examples

let tcb = new_mock_tcb_with_state(ThreadState::ThreadStateRunning);
assert_eq!(tcb.get_state(), ThreadState::ThreadStateRunning);
let cspace_0 = tcb.get_cspace(0);
assert_eq!(cspace_0.cteMDBNode.words[0], 0);
assert_eq!(cspace_0.cteMDBNode.words[1], 0);
assert_eq!(cspace_0.cap.words[0], 0);
assert_eq!(cspace_0.cap.words[1], 0);

pub fn init(&mut self)

Initialize the TCB

pub fn get_cspace_mut_ref(&mut self, i: usize) -> &'static mut cte_t

Get i th cspace of the TCB, mutable reference

pub fn get_state(&self) -> ThreadState

Get the current state of the TCB

pub fn is_stopped(&self) -> bool

Check if the TCB is stopped by checking the state

Examples

let stopped_states = vec![
    ThreadState::ThreadStateInactive,
    ThreadState::ThreadStateBlockedOnReceive,
    ThreadState::ThreadStateBlockedOnSend,
    ThreadState::ThreadStateBlockedOnReply,
    ThreadState::ThreadStateBlockedOnNotification,
];

stopped_states.into_iter().for_each(|state| {
    let tcb = new_mock_tcb_with_state(state);
    assert_eq!(tcb.is_stopped(), true);
});

pub fn is_runnable(&self) -> bool

Check if the TCB is runnable by checking the state

Examples

let states = vec![
    ThreadState::ThreadStateRunning,
    ThreadState::ThreadStateRestart,
];

states.into_iter().for_each(|state| {
    let tcb = new_mock_tcb_with_state(state);
    assert_eq!(tcb.is_runnable(), true);
});

pub fn is_current(&self) -> bool

Check if the TCB is current by comparing the tcb pointer

Example

let tcb = &mut new_mock_tcb_with_state(ThreadState::ThreadStateRunning);
set_current_thread(tcb);
assert_eq!(tcb.is_current(), true);

pub fn set_mcp_priority(&mut self, mcp: usize)

Examples

let mut tcb = new_mock_tcb_with_state(ThreadState::ThreadStateRunning);
let target_mcps = vec![20, 40, 60];

target_mcps.into_iter().for_each(|target_mcp| {
    tcb.set_mcp_priority(target_mcp);

    assert_eq!(tcb.tcbMCP, target_mcp);
});

pub fn set_priority(&mut self, priority: usize)

Set the priority of the TCB, and reschedule if the thread is runnable and not current

pub fn bind_notification(&mut self, addr: pptr_t)

Bind the notification of the TCB

Arguments

addr - The address of the notification to bind.

Example

let tcb = &mut new_mock_tcb_with_state(ThreadState::ThreadStateRunning);
let target_ntf: Vec<pptr_t> = vec![0x40000, 0x400000, 0x4000000];

target_ntf.into_iter().for_each(|ntf_ptr| {
    tcb.bind_notification(ntf_ptr);

    assert_eq!(tcb.tcbBoundNotification, ntf_ptr);
});

pub fn unbind_notification(&mut self)

Unbind the notification of the TCB(just set the bound notification to 0)

Example

let tcb = &mut new_mock_tcb_with_state(ThreadState::ThreadStateRunning);
tcb.bind_notification(0x20);
assert_ne!(tcb.tcbBoundNotification, 0);

tcb.unbind_notification();
assert_eq!(tcb.tcbBoundNotification, 0);

pub fn set_domain(&mut self, dom: usize)

Set the domain of the TCB.

Example

let tcb = &mut new_mock_tcb_with_state(ThreadState::ThreadStateRunning);
let target_domain = 0;
assert_eq!(tcb.tcbState.get_tcb_queued(), 0);
tcb.set_domain(target_domain);
assert_eq!(tcb.domain, target_domain);
assert_eq!(tcb.is_runnable(), true);
assert_eq!(tcb.tcbState.get_tcb_queued(), 1);

let tcb = &mut new_mock_tcb_with_state(ThreadState::ThreadStateRunning);
let target_domain = 0;
set_current_scheduler_action(SchedulerAction_ResumeCurrentThread);
assert_eq!(tcb.tcbState.get_tcb_queued(), 0);
set_current_thread(tcb);
assert_eq!(tcb.is_current(), true);

tcb.set_domain(target_domain);
assert_eq!(tcb.domain, target_domain);
assert_eq!(tcb.tcbState.get_tcb_queued(), 1);
assert_eq!(get_ks_scheduler_action(), SchedulerAction_ChooseNewThread);

pub fn sched_enqueue(&mut self)

Enqueue the TCB to the scheduling queue

Example

let tcb = &mut new_mock_tcb_with_state(ThreadState::ThreadStateRunning);
let target_domain = 0;
let target_priority = 200;
tcb.domain = target_domain;
tcb.set_priority(target_priority);

tcb.sched_enqueue();

assert_eq!(tcb.domain, target_domain);
assert_eq!(tcb.tcbPriority, target_priority);
assert_eq!(tcb.tcbSchedPrev, 0);
assert_eq!(tcb.tcbSchedNext, 0);
assert_eq!(tcb.tcbState.get_tcb_queued(), 1);

let idx = ready_queues_index(target_domain, target_priority);
let queue = tcb.get_sched_queue(idx);
assert_eq!(queue.head, tcb.get_ptr());
assert_eq!(queue.tail, tcb.get_ptr());

pub fn get_sched_queue(&mut self, index: usize) -> &'static mut tcb_queue_t

Get the scheduling queue by index from ksReadyQueues

pub fn get_cpu(&self) -> usize

Get the CPU of the TCB, 0 if not in SMP, tcbAffinity if in SMP

pub fn sched_dequeue(&mut self)

Dequeue the TCB from the scheduling queue

Example

let mut tcbs = vec![
new_mock_tcb_with_state(ThreadState::ThreadStateRunning),
new_mock_tcb_with_state(ThreadState::ThreadStateRunning),
new_mock_tcb_with_state(ThreadState::ThreadStateRunning),
];

let target_domain = 0;
let target_priority = 200;

let queue = tcbs[0].get_sched_queue(ready_queues_index(target_domain, target_priority));
assert!(queue.empty());

tcbs.iter_mut().for_each(|tcb| {
    tcb.domain = target_domain;
    tcb.set_priority(target_priority);
    tcb.sched_enqueue();
});
assert!(queue.empty() == false);

tcbs.iter_mut().for_each(|tcb| {
    tcb.sched_dequeue();
});
assert!(queue.empty());

pub fn sched_append(&mut self)

Append the TCB to the scheduling queue tail

Note

This function is as same as sched_enqueue, but it is used for the EP queue

pub fn set_vm_root(&self) -> Result<(), lookup_fault_t>

Set the VM root of the TCB

Example

let tcb = &mut new_mock_tcb_with_state(ThreadState::ThreadStateRunning);
let ret = tcb.set_vm_root();
assert_eq!(ret.is_ok(), true);

pub fn switch_to_this(&mut self)

Switch to the TCB(set current thread to self)

Example

let tcb = &mut new_mock_tcb_with_state(ThreadState::ThreadStateRunning);
tcb.sched_enqueue();
assert_ne!(get_currenct_thread().get_ptr(), tcb.get_ptr());

tcb.switch_to_this();
assert_eq!(get_currenct_thread().get_ptr(), tcb.get_ptr());

pub fn get_ptr(&self) -> pptr_t

Get the pointer of the TCB

Returns

The raw pointer of the TCB

pub fn lookup_slot(&self, cap_ptr: usize) -> lookupSlot_raw_ret_t

Look up the slot of the TCB

Arguments

cap_ptr - The capability pointer to look up

Returns

The lookup result structure

Example

let tcb = &mut new_mock_tcb_with_state(ThreadState::ThreadStateRunning);
let slot: &mut cte_t = &mut cte_t::default();
let slot_ptr = slot.get_ptr();
let guard_bits = wordBits - 1;
let radix_bits = 1;
let level_bits = radix_bits + guard_bits;
let cap_ptr = 0;

assert_eq!(level_bits, wordBits);
assert!(guard_bits <= wordBits);

let capCnodeGuard = (cap_ptr >> ((wordBits - guard_bits) & MASK!(wordRadix))) & MASK!(guard_bits);
assert_eq!(capCnodeGuard, 0);

let ctable_slot = tcb.get_cspace_mut_ref(tcbCTable);
ctable_slot.cap = cap_t::new_cnode_cap(1, guard_bits, capCnodeGuard , slot_ptr);
let slot = tcb.lookup_slot(0);

assert_eq!(slot.status, exception_t::EXCEPTION_NONE);
assert_eq!(slot.slot as *const cte_t as usize, slot_ptr);

pub fn setup_reply_master(&mut self)

Setup the reply master of the TCB

Example

let mut tcb = &mut new_mock_tcb_with_state(ThreadState::ThreadStateRunning);
// 改变 tcb指针
if tcb.get_ptr() - tcb.get_cspace(tcbCTable).get_ptr() < 3 * size_of::<cte_t>() {
    let tcb_ptr = tcb.get_cspace(tcbCTable).get_ptr() + 3 * size_of::<cte_t>();
    tcb = unsafe { &mut *(tcb_ptr as *mut tcb_t) };
}

assert_eq!(
    tcb.get_ptr() - tcb.get_cspace(tcbCTable).get_ptr() >= 3 * size_of::<cte_t>(),
    true
);
tcb.get_cspace_mut_ref(tcbReply).cap = cap_t::new_null_cap();
tcb.setup_reply_master();

assert_eq!(tcb.get_cspace(tcbReply).cap.get_cap_type(), CapTag::CapReplyCap);
assert_eq!(tcb.get_cspace(tcbReply).cap.get_type(), CapTag::CapReplyCap as usize);

pub fn suspend(&mut self)

Susupend the TCB, set the state to ThreadStateInactive and dequeue from the scheduling queue

Example

let tcb = &mut new_mock_tcb_with_state(ThreadState::ThreadStateRunning);
tcb.sched_enqueue();
assert_eq!(tcb.get_state(), ThreadState::ThreadStateRunning);
assert_eq!(tcb.tcbState.get_tcb_queued(), 1);
tcb.suspend();
assert_eq!(tcb.get_state(), ThreadState::ThreadStateInactive);
assert_eq!(tcb.tcbState.get_tcb_queued(), 0);

pub fn restart(&mut self)

Restart the TCB, set the state to ThreadStateRestart and enqueue to the scheduling queue waiting for reschedule

Example

let tcb = &mut new_mock_tcb_with_state(ThreadState::ThreadStateBlockedOnSend);
assert_eq!(tcb.get_state(), ThreadState::ThreadStateBlockedOnSend);
assert!(tcb.tcbState.get_tcb_queued() == 0);

tcb.restart();

assert_eq!(tcb.get_state(), ThreadState::ThreadStateRestart);
assert_eq!(tcb.tcbState.get_tcb_queued(), 1);

pub fn setup_caller_cap(&mut self, sender: &mut Self, can_grant: bool)

Setup the caller cap of the TCB

Arguments

sender - The sender TCB

can_grant - If the cap can be granted

pub fn delete_caller_cap(&mut self)

Delete the caller cap of the TCB

Example

let tcb = &mut new_mock_tcb_with_state(ThreadState::ThreadStateRunning);
tcb.get_cspace_mut_ref(tcbCaller).cap = cap_t::new_reply_cap(0, 0, 0);
tcb.delete_caller_cap();
assert_eq!(
    tcb.get_cspace(tcbCaller).cap.get_cap_type(),
    CapTag::CapNullCap
);

pub fn lookup_ipc_buffer( &self, is_receiver: bool, ) -> Option<&'static seL4_IPCBuffer>

Look up the IPC buffer of the TCB

Arguments

is_receiver - If the TCB is receiver

Returns

The IPC buffer of the TCB

Example

let page_base: [u8; BIT!(seL4_PageBits)] = [0; BIT!(seL4_PageBits)]; // page
let tcb = &mut new_mock_tcb_with_state(ThreadState::ThreadStateRunning);
tcb.tcbIPCBuffer = 0x20;
let buffer_cap = cap_t::new_frame_cap(
    0,
    page_base.as_ptr() as usize,
    seL4_PageBits,
    vm_rights_t::VMReadWrite as usize,
    0,
    0,
);

let mock_buffer =
    convert_to_mut_type_ref::<seL4_IPCBuffer>(page_base.as_ptr() as usize + 0x20);
mock_buffer.tag = 0x888;
mock_buffer.msg = [0x200; seL4_MsgMaxLength];
mock_buffer.userData = 0x400;
mock_buffer.caps_or_badges = [0x600; seL4_MsgMaxExtraCaps];
mock_buffer.receiveCNode = 0x800;
mock_buffer.receiveIndex = 0x1000;
mock_buffer.receiveDepth = 0x2000;

tcb.get_cspace_mut_ref(tcbBuffer).cap = buffer_cap;

let res = tcb.lookup_ipc_buffer(false);
assert_eq!(res.is_some(), true);
let res = res.unwrap();
assert_eq!(res.tag, 0x888);
assert_eq!(res.msg, [0x200; seL4_MsgMaxLength]);
assert_eq!(res.userData, 0x400);
assert_eq!(res.caps_or_badges, [0x600; seL4_MsgMaxExtraCaps]);
assert_eq!(res.receiveCNode, 0x800);
assert_eq!(res.receiveIndex, 0x1000);
assert_eq!(res.receiveDepth, 0x2000);

pub fn lookup_extra_caps( &self, res: &mut [pptr_t; 3], ) -> Result<(), seL4_Fault_t>

Look up the extra caps of the TCB

Arguments

res - The result array to store the extra caps

Returns

The result of the lookup represented by seL4_Fault_t

Example

let page_base: [u8; BIT!(seL4_PageBits)] = [0; BIT!(seL4_PageBits)]; // page
let tcb = &mut new_mock_tcb_with_state(ThreadState::ThreadStateRunning);

// ipc buffer build
{
    tcb.tcbIPCBuffer = 0x20;
    let buffer_cap = cap_t::new_frame_cap(
        0,
        page_base.as_ptr() as usize,
        seL4_PageBits,
        vm_rights_t::VMReadWrite as usize,
        0,
        0,
    );

    let mock_buffer =
        convert_to_mut_type_ref::<seL4_IPCBuffer>(page_base.as_ptr() as usize + 0x20);
    mock_buffer.tag = 0x888;
    mock_buffer.msg = [0x200; seL4_MsgMaxLength];
    mock_buffer.userData = 0x400;
    mock_buffer.caps_or_badges = [0; seL4_MsgMaxExtraCaps];
    mock_buffer.receiveCNode = 0x800;
    mock_buffer.receiveIndex = 0x1000;
    mock_buffer.receiveDepth = 0x2000;

    tcb.get_cspace_mut_ref(tcbBuffer).cap = buffer_cap;
}

// slot build
let slot: &mut cte_t = &mut cte_t::default();
{
    let slot_ptr = slot.get_ptr();
    let guard_bits = wordBits - 1;
    let radix_bits = 1;
    let level_bits = radix_bits + guard_bits;
    let cap_ptr = 0;

    assert_eq!(level_bits, wordBits);
    assert!(guard_bits <= wordBits);

    let capCnodeGuard =
        (cap_ptr >> ((wordBits - guard_bits) & MASK!(wordRadix))) & MASK!(guard_bits);
    assert_eq!(capCnodeGuard, 0);

    let ctable_slot = tcb.get_cspace_mut_ref(tcbCTable);
    ctable_slot.cap = cap_t::new_cnode_cap(1, guard_bits, capCnodeGuard, slot_ptr);
}

// msg info build
{
    tcb.tcbArch.set_register(ArchReg::MsgInfo, 1 << 7);
}

let res: &mut [pptr_t; 3] = &mut [0; seL4_MsgMaxExtraCaps];
let result = tcb.lookup_extra_caps(res);
assert_eq!(result.is_ok(), true);
assert_eq!(res[0], slot.get_ptr());
assert_eq!(res[1], 0);
assert_eq!(res[2], 0);

pub fn lookup_extra_caps_with_buf( &self, res: &mut [pptr_t; 3], buf: Option<&seL4_IPCBuffer>, ) -> Result<(), seL4_Fault_t>

Look up the extra caps of the TCB with IPC buffer

Arguments

res - The result array to store the extra caps

buf - The IPC buffer to look up

Returns

The result of the lookup represented by seL4_Fault_t

Example

let page_base: [u8; BIT!(seL4_PageBits)] = [0; BIT!(seL4_PageBits)]; // page
let tcb = &mut new_mock_tcb_with_state(ThreadState::ThreadStateRunning);

// ipc buffer build

tcb.tcbIPCBuffer = 0x20;
let buffer_cap = cap_t::new_frame_cap(
    0,
    page_base.as_ptr() as usize,
    seL4_PageBits,
    vm_rights_t::VMReadWrite as usize,
    0,
    0,
);

let mock_buffer =
    convert_to_mut_type_ref::<seL4_IPCBuffer>(page_base.as_ptr() as usize + 0x20);
mock_buffer.tag = 0x888;
mock_buffer.msg = [0x200; seL4_MsgMaxLength];
mock_buffer.userData = 0x400;
mock_buffer.caps_or_badges = [0; seL4_MsgMaxExtraCaps];
mock_buffer.receiveCNode = 0x800;
mock_buffer.receiveIndex = 0x1000;
mock_buffer.receiveDepth = 0x2000;
drop(mock_buffer);

tcb.get_cspace_mut_ref(tcbBuffer).cap = buffer_cap;

// slot build
let slot: &mut cte_t = &mut cte_t::default();
{
    let slot_ptr = slot.get_ptr();
    let guard_bits = wordBits - 1;
    let radix_bits = 1;
    let level_bits = radix_bits + guard_bits;
    let cap_ptr = 0;

    assert_eq!(level_bits, wordBits);
    assert!(guard_bits <= wordBits);

    let capCnodeGuard =
        (cap_ptr >> ((wordBits - guard_bits) & MASK!(wordRadix))) & MASK!(guard_bits);
    assert_eq!(capCnodeGuard, 0);

    let ctable_slot = tcb.get_cspace_mut_ref(tcbCTable);
    ctable_slot.cap = cap_t::new_cnode_cap(1, guard_bits, capCnodeGuard, slot_ptr);
}

// msg info build
{
    tcb.tcbArch.set_register(ArchReg::MsgInfo, 1 << 7);
}

let res: &mut [pptr_t; 3] = &mut [0; seL4_MsgMaxExtraCaps];
let mock_buffer = convert_to_type_ref::<seL4_IPCBuffer>(page_base.as_ptr() as usize + 0x20);
let result = tcb.lookup_extra_caps_with_buf(res, Some(mock_buffer));
assert_eq!(result.is_ok(), true);

assert_eq!(res[0], slot.get_ptr());
assert_eq!(res[1], 0);
assert_eq!(res[2], 0);

pub fn lookup_mut_ipc_buffer( &mut self, is_receiver: bool, ) -> Option<&'static mut seL4_IPCBuffer>

As same as lookup_ipc_buffer, but the result is mutable reference

pub fn set_mr(&mut self, offset: usize, reg: usize) -> usize

Set the message info register of the TCB

Arguments

offset - The offset of the message info register, if the offset is larger than n_msgRegisters, set to the IPC buffer

reg - The value to set

Returns

The next offset

pub fn set_lookup_fault_mrs( &mut self, offset: usize, fault: &lookup_fault_t, ) -> usize

Set the lookup fault to the msg registers of the TCB

Arguments

offset - The offset of the lookup fault

fault - The lookup fault to set

Returns

The next offset

pub fn get_receive_slot(&mut self) -> Option<&'static mut cte_t>

Get the receive slot of the TCB

Returns

The mutable ref of receive slot of the TCB

pub fn copy_mrs(&self, receiver: &mut tcb_t, length: usize) -> usize

Copy the message registers and ipc buffer(if valid) of the TCB to the receiver

Arguments

receiver - The receiver TCB

length - The length of the message registers to copy

Returns

The number of registers(contains ipc buffer) copied

pub fn copy_fault_mrs(&self, receiver: &mut Self, id: usize, length: usize)

Copy the falut messages and ipc buffer(if valid) of the TCB to the receiver

Arguments

receiver - The receiver TCB

id - The fault message id

length - The length of the message registers to copy

pub fn copy_fault_mrs_for_reply( &self, receiver: &mut Self, id: usize, length: usize, )

Copy the falut messages for reply and ipc buffer(if valid) of the TCB to the receiver for reply

Arguments

receiver - The receiver TCB

id - The fault message id

length - The length of the message registers to copy

pub fn copy_syscall_fault_mrs(&self, receiver: &mut Self)

Copy the syscall fault messages of the TCB to the receiver

pub fn copy_exeception_fault_mrs(&self, receiver: &mut Self)

Copy the exception fault messages of the TCB to the receiver

pub fn set_fault_mrs(&self, receiver: &mut Self) -> usize

Set the fault message registers of the TCB to the receiver

Arguments

receiver - The receiver TCB